+++ src/main/java/br/com/invistaix/InvistaIX/repository/ImpostoRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-    public interface ImpostoRepository extends JpaRepository<ImpostoModel, Integer> {
+    public interface ImpostoRepository extends JpaRepository<ImpostoModel, Long> {
     List<ImpostoModel> findByidimovel(Long idimovel);
     Optional<ImpostoModel> findByIdimovelAndDataBetween(Long idimovel, LocalDate inicioAno, LocalDate fimAno);
     @Query("SELECT COUNT(i) > 0 FROM ImpostoModel i WHERE i.idimovel = :idimovel AND EXTRACT(YEAR FROM i.data) = :ano")
